+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/DiagnosticTestGenerated.java
@@ -1923,6 +1923,12 @@ public class DiagnosticTestGenerated extends AbstractDiagnosticTest {
     }
 
     @Test
+    @TestMetadata("AssistedFactorySuspendMultibindings.kt")
+    public void testAssistedFactorySuspendMultibindings() {
+      run("AssistedFactorySuspendMultibindings.kt");
+    }
+
+    @Test
     @TestMetadata("AssistedFactorySuspendRequired.kt")
     public void testAssistedFactorySuspendRequired() {
       run("AssistedFactorySuspendRequired.kt");