+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/JsDiagnosticTestGenerated.java
@@ -1719,6 +1719,12 @@ public class JsDiagnosticTestGenerated extends AbstractJsDiagnosticTest {
     }
 
     @Test
+    @TestMetadata("AssistedFactorySuspendMultibindings.kt")
+    public void testAssistedFactorySuspendMultibindings() {
+      run("AssistedFactorySuspendMultibindings.kt");
+    }
+
+    @Test
     @TestMetadata("AssistedFactorySuspendRequired.kt")
     public void testAssistedFactorySuspendRequired() {
       run("AssistedFactorySuspendRequired.kt");